+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## main
+
+- better worksapce error state handling
+
 ## 9.0.16 2026/02/01
 
 - add multiple defs support to program window